Feature/bdn 215 bidding network (#84)
* BDN-215. Bidding. Separated Bidding/AdNetwork logic

* BDN-215. Bidding. Models. List Serializer. Parser.

* BDN-215. Bidding. Moved Geo to Device.

* BDN-215. Bidding. Tests

* BDN-220. Added Reg object for all requests. Added Test object.

* BDN-215. Bidding. Auction impl.

* BDN-215. Bidding. Added fields round_id to bidding request. Removed unused response-fields.

* BDN-215. Bidding. Renamed AdSourceType class to AdLoadingType.

* BDN-215. Bidding. Added auction_id, auction_configuration_id. Ktlint

* BDN-215. Bidding. Simplified AdAuctionParams. Updated bidding auction models.

* BDN-224. Added auction start/finish time

* BDN-215. Bidding. Updated bidding auction and stats models.

* BDN-215. Bidding. Extracted ExecuteRoundUseCase

* BDN-215. Bidding. Added interstitial/rewarded objects to imp requests.

* BDN-215. Bidding. Refactoring.

* BDN-215. Bidding. Refactoring.

* BDN-236. Added bid.status

* BDN-215. Bidding. Updated auction bidding logic. Added stat tracking. Removed WinLoss notification for BidMachine.

* BDN-215. Bidding. Tests

* BDN-215. Bidding. Added auction_configuration_id to /loss-request

* BDN-215. Bidding. Fixed /loss-request after show

* BDN-215. Bidding. Changed data types: accuracy is Int, lastfix in sec.

* BDN-215. Bidding. Removed unused

* BDN-215. Fixed disappearing of BannerView Compose

+++ bidon/src/main/java/org/bidon/sdk/databinders/location/LocationDataSource.kt
@@ -17,7 +17,7 @@ internal interface LocationDataSource : DataSource {
      */
     fun getLongitude(): Double?
 
-    fun getAccuracy(): Float?
+    fun getAccuracy(): Int?
     fun getLastFix(): Long?
 
     /**
